finish-use-sort
Finished the counter component, with a custom hook removing the complexity of the component, leaving only the jsx.

Fixed table warning with fragment.
