+++ src/components/table/SortableTable.js
@@ -1,33 +1,16 @@
-import { useState } from "react";
+import useSort from "../../hooks/use-sort";
 import Table from "./Table";
 import { GoTriangleDown, GoTriangleUp } from "react-icons/go";
 
 function SortableTable(props) {
 
-  const [sortOrder, setSortOrder] = useState(null);
-  const [sortBy, setSortBy] = useState(null);
-
   const { config, data } = props;
 
-  const handleClick = (label) => {
-
-    if (sortBy && label !== sortBy) {
-      setSortOrder('asc');
-      setSortBy(label)
-      return;
-    }
-
-    if (sortOrder === null) {
-      setSortOrder('asc');
-      setSortBy(label);
-    } else if (sortOrder === 'asc') {
-      setSortOrder('desc');
-      setSortBy(label);
-    } else if (sortOrder === 'desc') {
-      setSortOrder(null);
-      setSortBy(null);
-    }
-  }
+  const {
+    sortBy,
+    sortOrder,
+    sortedData,
+    setSortColumn } = useSort(data, config);
 
   const updatedConfig = config.map((column) => {
     if (!column.sortValue) {
@@ -35,7 +18,9 @@ function SortableTable(props) {
     }
     return {
       ...column,
-      header: () => <th className="cursor-pointer hover:bg-gray-100" onClick={() => handleClick(column.label)} >
+      header: () => <th className="cursor-pointer hover:bg-gray-100"
+        onClick={() => setSortColumn(column.label)}
+      >
         <div className="flex items-center">
           {getIcons(column.label, sortBy, sortOrder)}
           {column.label}
@@ -43,24 +28,6 @@ function SortableTable(props) {
       </th>
     }
   });
-
-  let sortedData = data;
-
-  if (sortOrder && sortBy) {
-    const { sortValue } = config.find(column => column.label === sortBy);
-    sortedData = [...data].sort((a, b) => {
-      const valueA = sortValue(a);
-      const valueB = sortValue(b);
-
-      const reverseOrder = sortOrder === 'asc' ? 1 : -1;
-
-      if (typeof valueA === 'string') {
-        return valueA.localeCompare(valueB) * reverseOrder;
-      } else {
-        return (valueA - valueB) * reverseOrder;
-      }
-    });
-  }
 
   return (
     <Table {...props} data={sortedData} config={updatedConfig} />
